financial_pipeline/requirements.txt
pandas
numpy
requests
pyyaml
python-dotenv
streamlit
matplotlib
seaborn
plotly
pytest
pyarrow
fastparquet
